sql related changes
changed the schema to include a path to the items image, created a notebook for running item management functions quickly and easily, not really part of the bot as much as just an item database development tool. id consider other fixes but this one is simple enough that im happy with it for now.